Fixed setup process and rearranged index.ts
